requirements.txt
asgiref==3.10.0
black==25.9.0
cfgv==3.4.0
click==8.3.0
coverage==7.11.1
distlib==0.4.0
Django==5.2.8
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
djangorestframework==3.16.1
filelock==3.20.0
flake8==7.3.0
identify==2.6.15
iniconfig==2.3.0
isort==7.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-decouple==3.8
pytokens==0.3.0
PyYAML==6.0.3
sqlparse==0.5.3
virtualenv==20.35.4
